refactor: extract Onboarding page components into separate files
Extracted `SetupIcon` and `IconLabel` into dedicated component files to improve
readability and maintainability of the Onboarding wizard page (`src/ui/next/src/app/onboarding/page.tsx`).
Also fixed broken API import route.
