+++ src/ui/next/src/app/api/ui/dashboard/daily-work/action/route.ts
@@ -1,4 +1,4 @@
-import { proxyBackendPost } from "../../backendProxy";
+import { proxyBackendPost } from "../../../backendProxy";
 import { NextResponse } from "next/server";
 
 export async function POST(req: Request) {
